+++ application/views/User/History.php
@@ -35,7 +35,7 @@
 </table>
 
 <nav aria-label="Page navigation" style="text-align:center">
-	<ul id="history-pagination" class="pagination">
+	<ul class="pagination">
 		<li class="page-item <?=($currentPage == 1 ? 'disabled' : '')?>">
 			<a class="page-link" href="<?=base_url("user/history/".($currentPage - 1))?>" aria-label="Previous">
 				<span aria-hidden="true">&laquo;</span>
@@ -60,5 +60,6 @@
 <script src="<?=asset_url()?>vendor/js/jquery.simplePagination.js" defer></script>
 <script>
 	const currentPagination = parseInt("<?=$currentPage?>");
-	const totalPagination = parseInt("<?=$totalPages?>");
+	const totalPagination   = parseInt("<?=$totalPages?>");
+	const titleID           = 0; //Hack
 </script>