+++ public/assets/js/pages/history.js
@@ -3,28 +3,13 @@ $(function(){
 	"use strict";
 	if(page !== 'history') { return false; }
 
-	$('#history-pagination').pagination({
+	$('.pagination').pagination({
 		currentPage    : currentPagination,
 
 		pages          : totalPagination,
 		displayedPages : 10,
 
-		hrefTextPrefix : '/user/history/',
-
-		selectOnClick  : false,
-
-		prevText       : '&laquo;',
-		nextText       : '&raquo;'
-	});
-
-	if(typeof titleID === 'undefined') { const titleID = 0; } //handle possible error if titleID isn't set
-	$('#title-history-pagination').pagination({
-		currentPage    : currentPagination,
-
-		pages          : totalPagination,
-		displayedPages : 10,
-
-		hrefTextPrefix : '/history/'+titleID+'/',
+		hrefTextPrefix : (/user\/history/.test(location.pathname) ? '/user/history/' : '/history/'+titleID+'/'),
 
 		selectOnClick  : false,